Solved std::tolower warning

+++ src/ImageViewer.cpp
@@ -210,6 +210,6 @@ ImageViewerStatus ImageViewer::getStatus()
 bool ImageViewer::isSupportedImageType(std::experimental::filesystem::path file)
 {
     std::string extension = file.string().substr(file.string().find_last_of(".")+1,file.string().size());
-    std::transform(extension.begin(), extension.end(), extension.begin(), [](unsigned char c){ return std::tolower((unsigned int) c);});
+    std::transform(extension.begin(), extension.end(), extension.begin(), [](unsigned char c){return std::tolower(c);});
     return in<std::string>(extension,{"bmp","png","tga","jpg","jpeg","jfif","gif","psd","hdr","pic"});
 }
